Refactor test cases to use const constructors where applicable
- Updated instances of TilawaIllustratedState, TilawaBottomActionArea, and other widgets to use const constructors for improved performance.
- Ensured consistency in widget construction across various test files, including tilawa_form_screen_scaffold_test.dart and tilawa_async_content_test.dart.
- Made adjustments to assertions in tests to accommodate changes in widget instantiation.

+++ apps/tilawa/lib/features/home/presentation/screens/home_screen.dart
@@ -266,9 +266,10 @@ class _HomeScreenState extends State<HomeScreen> {
       return;
     }
 
-    final double clampedTarget = snapTarget
-        .clamp(position.minScrollExtent, position.maxScrollExtent)
-        .toDouble();
+    final double clampedTarget = snapTarget.clamp(
+      position.minScrollExtent,
+      position.maxScrollExtent,
+    );
     if ((position.pixels - clampedTarget).abs() <=
         HomeScreen._heroSnapTolerance) {
       return;
