+++ apps/tilawa/lib/features/home/presentation/widgets/home_hero_photo_theme.dart
@@ -87,7 +87,7 @@ abstract final class HomeHeroPhotoTheme {
             alpha: screenTokens.homeCollapsedHeaderShadowOpacity,
           ),
           offset: tokens.shadowOffsetSmall,
-          blurRadius: tokens.spaceSmall.toDouble(),
+          blurRadius: tokens.spaceSmall,
         ),
       ],
     );
