+++ apps/tilawa/lib/features/smart_khatma/domain/usecases/extend_khatma_plan_use_case.dart
@@ -18,7 +18,7 @@ final class ExtendKhatmaPlanUseCase {
         return const Right(null);
       }
       final today = now ?? DateTime.now();
-      final int extraDays = plan.missedDays(today).clamp(1, 30).toInt();
+      final int extraDays = plan.missedDays(today).clamp(1, 30);
       final KhatmaPlan updated = plan.copyWith(
         durationDays: plan.durationDays + extraDays,
       );
